move over the oddysee calls

+++ Source/AliveLibAE/Electrocute.cpp
@@ -114,7 +114,7 @@ private:
         if (!field_CE_bDone)
         {
             // TODO: FIX ME - abstraction break, the x value is used as an offset as to how much to overwrite, the width isn't isn't the pal depth in this case
-            IRenderer::GetRenderer()->PalSetData(IRenderer::PalRecord{field_20_pal_xy.field_2_y, field_20_pal_xy.field_0_x + field_C8_pal_x_index, field_CA_pal_w}, reinterpret_cast<BYTE*>(&field_B8_palBuffer[0]));
+            IRenderer::GetRenderer()->PalSetData(IRenderer::PalRecord{field_20_pal_xy.field_2_y, static_cast<short>(field_20_pal_xy.field_0_x + field_C8_pal_x_index), field_CA_pal_w}, reinterpret_cast<BYTE*>(&field_B8_palBuffer[0]));
         }
     }
